+++ rust-src/id/Cargo.toml
@@ -23,9 +23,6 @@ pairing = "0.15"
 derive_more = "0.99"
 thiserror = "1.0"
 
-# only for printing account addresses
-base58check = "0.1"
-
 [dependencies.eddsa_ed25519]
 path = "../eddsa_ed25519"
 version = "0"
